class renaming and file breakout
