+++ apps/vaporgui/ImportTab.h
@@ -3,14 +3,14 @@
 #include "Updatable.h"
 #include <QWidget>
 
-class PImportData;
+class PImportDataWidget;
 class MainForm;
 
 class ImportTab : public QWidget, public Updatable {
     Q_OBJECT
 
     VAPoR::ControlExec *_ce;
-    PImportData *_importer;
+    PImportDataWidget *_importer;
 
 public:
     ImportTab(VAPoR::ControlExec *ce, MainForm *mf);
