+++ apps/vaporgui/PImportDataLineItem.h
@@ -1,0 +1,29 @@
+#pragma once
+
+#include "PLineItem.h"
+
+namespace VAPoR {
+    class ControlExec;
+}
+class MainForm;
+class VHBoxWidget;
+class VLabel;
+class PButton;
+
+class PImportDataLineItem : public PLineItem {
+    Q_OBJECT
+
+    VAPoR::ControlExec *_ce;
+    MainForm *_mf;
+    VHBoxWidget *_hBox;
+    VLabel *_fileLabel;
+    PButton *_importButton;
+
+    void _importDataset();
+
+public:
+    PImportDataLineItem(VAPoR::ControlExec* ce, MainForm *mf);
+
+protected:
+    void updateGUI() const override;
+};
